+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 * Initial non-experimental version of SensorChanSpec with limited functionality
+* Basic jagged support for delayed load of data at native asset dimensions. 
 
 ### Fixed
 * Fix compatibility issue with datasets that do not contain an annotations list